Move nvhost-ctrl-gpu to new device API
